STYLE: Remove utilities.h. Move used code into the main class file.

+++ Modules/Segmentation/GrowCut/include/HeapNode.h
@@ -63,7 +63,7 @@ public:
     return N;
   }
   void
-  SetKeyValue(double n)
+  SetKeyValue(float n)
   {
     N = n;
   }
